+++ Api/Data/NodeInterface.php
@@ -16,6 +16,8 @@ interface NodeInterface
     const LEVEL = 'level';
     const TITLE = 'title';
     const TARGET = 'target';
+    const IMAGE = 'image';
+    const IMAGE_ALT_TEXT = 'image_alt_text';
     const CREATION_TIME = 'creation_time';
     const UPDATE_TIME = 'update_time';
     const IS_ACTIVE = 'is_active';
@@ -170,6 +172,36 @@ interface NodeInterface
      * @return $this
      */
     public function setTarget($target);
+
+    /**
+     * Get image
+     *
+     * @return string
+     */
+    public function getImage();
+
+    /**
+     * Set image
+     *
+     * @param string $image
+     * @return $this
+     */
+    public function setImage($image);
+
+    /**
+     * Get image alt text
+     *
+     * @return string
+     */
+    public function getImageAltText();
+
+    /**
+     * Set image alt text
+     *
+     * @param string $altText
+     * @return $this
+     */
+    public function setImageAltText($altText);
 
     /**
      * Get creation time