+++ view/adminhtml/web/vue/menu-type/category.vue
@@ -5,8 +5,7 @@
         :description="config.translation.categoryId"
         :item="item"
         :item-key="'content'"
-        :options="options"
-        :options-tree="optionsTree"
+        :options="config.fieldData.category.snowMenuAutoCompleteField.options"
         :config="config"
     />
 </template>
@@ -22,31 +21,6 @@
                 item: {
                     type: Object,
                     required: true
-                }
-            },
-            data() {
-                return {
-                    options: this.config.fieldData.category.snowMenuAutoCompleteField.options
-                }
-            },
-            computed: {
-                optionsTree() {
-                    const hashTable = {},
-                          optionsTree = [];
-
-                    this.options.forEach(item => hashTable[item.id] = {...item});
-                    this.options.forEach(item => {
-                        if (item.parent_id && hashTable[item.parent_id]) {
-                            hashTable[item.parent_id].children = [
-                                ...(hashTable[item.parent_id].children || []),
-                                hashTable[item.id]
-                            ];
-                        } else {
-                            optionsTree.push(hashTable[item.id]);
-                        }
-                    });
-
-                    return optionsTree;
                 }
             },
             template: template