+++ CHANGELOG.md
@@ -6,13 +6,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.14.0] - 2021-03-30
 ### Added
 - Node validation classes (#73442)
 - DB table columns listing method to menu and node resource models (#70197)
 - Menu YAML import/export feature (#70197)
 - Nodes custom templates documentation (#74567)
+- Menu node image field (#70218)
 - Treeselect for category selector (#69126)
 - GraphQL feature (#76409)
+- Treeselect for child category selector (#76417)
 
 ### Changed
 - Improve node product validation error message on menu save (#73442)